qtenv: Cast `QList::length()` and `QVector::size()` results to `int`
They got widened in Qt 6. Also, `QVector` is now an alias of
`QList`, and `length` is the same as `size` (or `count`).

+++ src/qtenv/inspectorutil.cc
@@ -205,7 +205,7 @@ QMenu *InspectorUtil::createInspectorContextMenu(QVector<cObject *> objects, Ins
         const int maxObjects = 20;
 
         // then create a submenu for each object
-        for (int i = 0; i < std::min(maxObjects, objects.size()); ++i) {
+        for (int i = 0; i < std::min(maxObjects, (int)objects.size()); ++i) {
             cObject *object = objects[i];
 
             const char *name = object->getFullName();
